Add command_handler
