+++ src/errors.rs
@@ -2,8 +2,6 @@ use thiserror::Error;
 
 #[derive(Error, Debug)]
 pub enum ParseError {
-    #[error("command '{0}' not found")]
-    CommandNotFound(String),
     #[error("missing argument: '{0}'")]
     MissingArgument(String),
 }
@@ -22,10 +20,6 @@ pub enum CommandError {
     ElasticsearchConnectionError(String),
     #[error("failed to get latest commit id ({0})")]
     FailedGetLatestCommitId(String),
-    #[error("failed to update package ({0})")]
-    FailedUpdatePackage(String),
-    #[error("failed to update all packages")]
-    FailedUpdateAllPackages(String),
     #[error("missing file ({0})")]
     MissingFile(String),
 }
